refactor: 클래스명 명사형으로 변경
